v6 AWS
Currently working on a couple of aws projects such as machine learning , and a django site that can be accessed online. ML is a big plus, since I currently lack a physical server, but using aws will further increase my skill set in using cloud computing to help aid in the creation of artifical intelligence!